fix: create temp files next to original files
Due to a [quirk](https://docs.rs/tempfile/3.4.0/tempfile/struct.NamedTempFile.html#method.persist) in the `tempfile` crate
temporary files can't be persisted across filesystems. This change fixes
the issue by creating the temporary file next to the original file.

I'm not sure if there's an easy way to create unit test for this, which
isn't the nicest. But hey, it'll do for now.

Fixes #65